Copy directory recursion fix
Changed the way the `Copy::createDelegate` method handles copying when the source file is actually a folder. Made this recurse back into the `create` method for now, passing the folder as a glob pattern. This is less performant than copying the entire folder at once, but copying the entire folder produces incorrect file paths when the destination folder already exists and copying file-by-file seems the best way to reconcile this so that any pre-existing non-composer-installed files in that folder can remain intact.